Implement a text mode for non-wordlist text files
Users can now supply any kind of plain text file to type its
contents exactly; as opposed to the shuffled word approach of the
wordlist mode.

One gotcha is that since the typing stage is of width 80, the
user-supplied text file should be word wrapped to 80 columns or it
will break the layout.